app/pyproject.toml
[tool.poetry]
name = "lab-kaggle-hub"
version = "0.1.0"
description = "Research-lab Kaggle experiment hub (walking skeleton)"
authors = ["you <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.5"}
pydantic = "^2.7.0"
psycopg2-binary = "^2.9.9"
SQLAlchemy = "^2.0.32"
python-dotenv = "^1.0.1"
boto3 = "^1.34.0"
jinja2 = "^3.1.6"

[tool.poetry.group.dev.dependencies]
httpx = "^0.27.0"
pytest = "^8.2.0"

[tool.poetry.scripts]
api = "app.api.main:run"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
